fix: 因同名类导致boot-assembly无法正常启动 #2397 调整同名类
