+++ scripts/bkenv.properties
@@ -28,6 +28,8 @@ BKCI_HTTP_PORT=80
 HTTP_SCHEMA=http
 # 如果有启用HTTPS才配置有效
 BKCI_HTTPS_PORT=443
+# 单体微服务端口
+BKCI__API_PORT=21900
 # 各微服务的HTTP端口 application-{service}.yml {service}.sh
 BKCI_NOTIFY_API_PORT=21911
 BKCI_PROJECT_API_PORT=21912
