+++ scripts/generate_boot.sh
@@ -16,7 +16,7 @@ fi
 for service_name in `awk -F':boot-' '/:boot-/ && !/assembly"/ {print $2}' ../src/backend/ci/settings.gradle | tr -d '"' | sed '$aagentless'`;do
   # convert lower to upper
   SERVICE_NAME=${service_name^^}
-  
+
   tpl_name="$service_name#boot-$service_name.sh"
   tpl_file="$tpl_dir/$tpl_name"
   if [ -s "$tpl_file" ];then
@@ -28,3 +28,14 @@ for service_name in `awk -F':boot-' '/:boot-/ && !/assembly"/ {print $2}' ../src
       -e "s/BKCISERVICE/$SERVICE_NAME/g" \
         "$tpl_tpl" > "$tpl_file"
 done
+
+# assembly ASSEMBLY
+tpl_name="assembly#boot-assembly.sh"
+tpl_file="$tpl_dir/$tpl_name"
+
+if [ -s "$tpl_file" ];then
+  echo "  SKIP: target file exist: $tpl_file"
+else
+  echo "  generating $tpl_file"
+  sed -e "s/bkciservice/assembly/g" -e "s/BKCISERVICE/ASSEMBLY/g" "$tpl_tpl" > "$tpl_file"
+fi
