+++ src/backend/ci/core/misc/biz-misc/src/main/kotlin/com/tencent/devops/misc/dao/EnvironmentNodeDao.kt
@@ -34,7 +34,7 @@ import org.jooq.DSLContext
 import org.springframework.stereotype.Repository
 
 @Repository
-class NodeDao {
+class EnvironmentNodeDao {
     fun get(dslContext: DSLContext, projectId: String, nodeId: Long): TNodeRecord? {
         with(TNode.T_NODE) {
             return dslContext.selectFrom(this)
