+++ src/backend/ci/core/process/api-process/src/main/kotlin/com/tencent/devops/process/api/builds/BuildRepositoryMaterialResource.kt
@@ -24,48 +24,42 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.artifactory.api.user
+package com.tencent.devops.process.api.builds
 
-import com.tencent.devops.common.api.auth.AUTH_HEADER_USER_ID
-import com.tencent.devops.common.api.auth.AUTH_HEADER_USER_ID_DEFAULT_VALUE
+import com.tencent.devops.common.api.auth.AUTH_HEADER_DEVOPS_BUILD_ID
+import com.tencent.devops.common.api.auth.AUTH_HEADER_DEVOPS_PIPELINE_ID
+import com.tencent.devops.common.api.auth.AUTH_HEADER_DEVOPS_PROJECT_ID
+import com.tencent.devops.common.api.pojo.Result
+import com.tencent.devops.process.pojo.PipelineBuildMaterial
 import io.swagger.annotations.Api
 import io.swagger.annotations.ApiOperation
 import io.swagger.annotations.ApiParam
 import javax.ws.rs.Consumes
-import javax.ws.rs.GET
 import javax.ws.rs.HeaderParam
+import javax.ws.rs.POST
 import javax.ws.rs.Path
-import javax.ws.rs.PathParam
 import javax.ws.rs.Produces
 import javax.ws.rs.core.MediaType
 
-@Api(tags = ["USER_REPORT"], description = "版本仓库-报告目录")
-@Path("/user/reports")
+@Api(tags = ["BUILD_REPOSITORY"], description = "构建-构建物料")
+@Path("/build/repository")
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
-interface UserReportResource {
-    @ApiOperation("获取有权限目录列表")
-    // @Path("/projects/{projectId}/pipelines/{pipelineId}/builds/{buildId}/elements/{elementId}/paths/{path: .*}")
-    @Path("/{projectId}/{pipelineId}/{buildId}/{elementId}/{path: .*}")
-    @GET
-    fun get(
-        @ApiParam("用户ID", required = true, defaultValue = AUTH_HEADER_USER_ID_DEFAULT_VALUE)
-        @HeaderParam(AUTH_HEADER_USER_ID)
-        userId: String,
-        @ApiParam("项目ID", required = true)
-        @PathParam("projectId")
-        projectId: String,
-        @ApiParam("流水线ID", required = true)
-        @PathParam("pipelineId")
-        pipelineId: String,
-        @ApiParam("构建ID", required = true)
-        @PathParam("buildId")
+interface BuildRepositoryMaterialResource {
+    @ApiOperation("保存构建物料信息")
+    @Path("/saveBuildMaterial")
+    @POST
+    fun saveBuildMaterial(
+        @ApiParam(value = "构建ID", required = true)
+        @HeaderParam(AUTH_HEADER_DEVOPS_BUILD_ID)
         buildId: String,
-        @ApiParam("原子ID", required = true)
-        @PathParam("elementId")
-        elementId: String,
-        @ApiParam("相对路径", required = true)
-        @PathParam("path")
-        path: String
-    )
+        @ApiParam(value = "项目ID", required = true)
+        @HeaderParam(AUTH_HEADER_DEVOPS_PROJECT_ID)
+        projectId: String,
+        @ApiParam(value = "流水线ID", required = true)
+        @HeaderParam(AUTH_HEADER_DEVOPS_PIPELINE_ID)
+        pipelineId: String,
+        @ApiParam(value = "物料列表", required = true)
+        pipelineBuildMaterials: List<PipelineBuildMaterial>
+    ): Result<Int>
 }
