+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/api/builds/BuildRepositoryMaterialResourceImpl.kt
@@ -24,15 +24,34 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.misc
+package com.tencent.devops.process.api.builds
 
-import com.tencent.devops.misc.service.SamplePipelineHistoryDataClearServiceImpl
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
+import com.tencent.devops.common.api.exception.ParamBlankException
+import com.tencent.devops.common.api.pojo.Result
+import com.tencent.devops.common.web.RestResource
+import com.tencent.devops.process.pojo.PipelineBuildMaterial
+import com.tencent.devops.process.service.PipelineBuildMaterialService
+import org.apache.commons.lang.StringUtils
+import org.springframework.beans.factory.annotation.Autowired
 
-@Configuration
-class SampleServiceConfig {
+@RestResource
+class BuildRepositoryMaterialResourceImpl @Autowired constructor(
+    private val pipelineBuildMaterialService: PipelineBuildMaterialService
+) : BuildRepositoryMaterialResource {
+    override fun saveBuildMaterial(
+        buildId: String,
+        projectId: String,
+        pipelineId: String,
+        pipelineBuildMaterials: List<PipelineBuildMaterial>
+    ): Result<Int> {
+        checkParam(buildId = buildId, projectId = projectId, pipelineId = pipelineId)
+        return Result(pipelineBuildMaterialService.saveBuildMaterial(buildId, projectId, pipelineId, pipelineBuildMaterials))
+    }
 
-    @Bean
-    fun pipelineHistoryDataClearService() = SamplePipelineHistoryDataClearServiceImpl()
+    fun checkParam(buildId: String, projectId: String, pipelineId: String) {
+        if (StringUtils.isBlank(buildId))
+            throw ParamBlankException("build Id is null or blank")
+        if (StringUtils.isBlank(pipelineId))
+            throw ParamBlankException("pipeline Id is null or blank")
+    }
 }
