+++ src/backend/ci/core/store/biz-store-sample/src/main/kotlin/com/tencent/devops/store/SampleStoreServiceConfig.kt
@@ -43,7 +43,7 @@ import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 
 @Configuration
-class SampleServiceConfig {
+class SampleStoreServiceConfig {
 
     @Bean
     fun containerService() = SampleContainerServiceImpl()
